+++ src/Accounts/Accounts/ChangeLog.md
@@ -21,6 +21,7 @@
 
 ## Version 1.2.1
 * Release with correct version of Authentication
+* Enable MSI Authentication in Azure Functions and WebApps
 
 ## Version 1.2.0
 * Add interactive and username/password authentication for Windows PowerShell 5.1 only